# openapi/spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/libraries/{pluginId}/{libraryId}": {
      "get": {
        "operationId": "getLibrary",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "libraryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginLibrary"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "Libraries"
        ]
      }
    },
    "/libraries/{pluginId}/{libraryId}/search": {
      "post": {
        "operationId": "searchLibrary",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "libraryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibrarySearchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryFindResponse"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "Libraries"
        ]
      }
    },
...
    "/artists/{artistUuid}": {
      "get": {
        "operationId": "getArtist",
        "parameters": [
          {
            "name": "artistUuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artist"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "Artists"
        ]
      }
    },
    "/artists": {
      "post": {
        "operationId": "searchArtists",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtistsSearchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistsSearchResponse"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Artists"
        ]
      }
    },
    "/artists/{artistUuid}/urls": {
      "get": {
        "operationId": "getArtistExternalUrls",
        "parameters": [
          {
            "name": "artistUuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExternalUrl"
                  }
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "Artists"
        ]
      }
    },
...
    "/attribute-sources/order": {
      "post": {
        "operationId": "setAttributeSourceOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttributeSourceOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttributeSource"
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "AttributeSources"
        ]
      }
    },
...
    "/tracks/{pluginId}/{libraryId}/{trackId}": {
      "get": {
        "operationId": "getTrack",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "libraryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Track"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Tracks"
        ]
      }
    },
    "/tracks/{pluginId}/{libraryId}/{trackId}/identities": {
      "get": {
        "operationId": "getTrackIdentities",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "libraryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Identity"
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Tracks"
        ]
      }
    },
    "/tracks/{pluginId}/{libraryId}/{trackId}/audio": {
      "get": {
        "operationId": "streamAudio",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "libraryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Range",
            "required": false,
            "in": "header",
            "description": "Byte range for seeking (e.g., bytes=0-1024)",
            "schema": {
              "type": "string",
              "example": "bytes=0-1024"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the full audio file when no range is requested",
            "content": {
              "audio/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "206": {
            "description": "Partial content when a valid range is provided",
            "headers": {
              "Content-Range": {
                "description": "The range of bytes sent and the total size of the file",
                "schema": {
                  "type": "string",
                  "example": "bytes 0-99/3539613"
                }
              },
              "Accept-Ranges": {
                "description": "Indicates that the server supports range requests",
                "schema": {
                  "type": "string",
                  "example": "bytes"
                }
              },
              "Content-Length": {
                "description": "Size of the chunk being returned",
                "schema": {
                  "type": "number"
                }
              }
            },
            "content": {
              "audio/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Library or Track not found"
          },
          "416": {
            "description": "Requested Range Not Satisfiable"
          }
        },
        "tags": [
          "Tracks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pipe Bomb API",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "PluginLibrary": {
        "type": "object",
        "properties": {
          "pluginId": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "pluginId",
          "id",
          "name"
        ]
      },
      "LibrarySearchDto": {
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 30
          },
          "page": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "pageSize",
          "page"
        ]
      },
      "StringAttribute": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "string"
            ]
          },
          "pluginId": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "pluginId",
          "sourceId",
          "values"
        ]
      },
      "BooleanAttribute": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "boolean"
            ]
          },
          "pluginId": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "boolean"
            }
          }
        },
        "required": [
          "type",
          "pluginId",
          "sourceId",
          "values"
        ]
      },
      "DecimalAttribute": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "decimal"
            ]
          },
          "pluginId": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "type",
          "pluginId",
          "sourceId",
          "values"
        ]
      },
      "IntegerAttribute": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "integer"
            ]
          },
          "pluginId": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "type",
          "pluginId",
          "sourceId",
          "values"
        ]
      },
      "Resource": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          }
        },
        "required": [
          "uuid",
          "url",
          "extension",
          "sha256"
        ]
      },
      "BufferAttribute": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "buffer"
            ]
          },
          "pluginId": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Resource"
            }
          }
        },
        "required": [
          "type",
          "pluginId",
          "sourceId",
          "values"
        ]
      },
      "Identity": {
        "type": "object",
        "properties": {
          "pluginId": {
            "type": "string"
          },
          "identityId": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "ordinal": {
            "type": "integer"
          }
        },
        "required": [
          "pluginId",
          "identityId",
          "entityId",
          "value",
          "ordinal"
        ]
      },
      "Artist": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "attributes": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AttributeMap"
              }
            ],
            "nullable": true
          },
          "identities": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Identity"
            }
          },
          "tracks": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "uuid",
          "attributes",
          "identities",
          "tracks"
        ]
      },
      "TrackArtist": {
        "type": "object",
        "properties": {
          "artistUuid": {
            "type": "string"
          },
          "artist": {
            "$ref": "#/components/schemas/Artist"
          },
          "joinPhrase": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "artistUuid",
          "artist",
          "joinPhrase"
        ]
      },
      "Track": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "pluginId": {
            "type": "string"
          },
          "libraryId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "artists": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackArtist"
            }
          },
          "attributes": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AttributeMap"
              }
            ],
            "nullable": true
          },
          "identities": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Identity"
            }
          }
        },
        "required": [
          "id",
          "pluginId",
          "libraryId",
          "title",
          "artists",
          "attributes",
          "identities"
        ]
      },
      "LibraryFindResponse": {
        "type": "object",
        "properties": {
          "tracks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Track"
            }
          }
        },
        "required": [
          "tracks"
        ]
      },
...
      "ArtistsSearchDto": {
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 30
          },
          "page": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "pageSize",
          "page"
        ]
      },
      "ArtistsSearchResponse": {
        "type": "object",
        "properties": {
          "artists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Artist"
            }
          }
        },
        "required": [
          "artists"
        ]
      },
      "ExternalUrl": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "iconUrl": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "iconUrl",
          "name"
        ]
      },
      "AttributeSource": {
        "type": "object",
        "properties": {
          "pluginId": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "pluginId",
          "sourceId",
          "name"
        ]
      },
      "OrderedAttributeSourceDto": {
        "type": "object",
        "properties": {
          "pluginId": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          }
        },
        "required": [
          "pluginId",
          "sourceId"
        ]
      },
      "AttributeSourceOrderDto": {
        "type": "object",
        "properties": {
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderedAttributeSourceDto"
            }
          }
        },
        "required": [
          "sources"
        ]
      },
...
      "AttributeMap": {
        "type": "object",
        "additionalProperties": {
          "oneOf": [
            {
              "$ref": "#/components/schemas/StringAttribute"
            },
            {
              "$ref": "#/components/schemas/BooleanAttribute"
            },
            {
              "$ref": "#/components/schemas/IntegerAttribute"
            },
            {
              "$ref": "#/components/schemas/DecimalAttribute"
            },
            {
              "$ref": "#/components/schemas/BufferAttribute"
            }
          ],
          "discriminator": {
            "propertyName": "type",
            "mapping": {
              "string": "#/components/schemas/StringAttribute",
              "boolean": "#/components/schemas/BooleanAttribute",
              "integer": "#/components/schemas/IntegerAttribute",
              "decimal": "#/components/schemas/DecimalAttribute",
              "buffer": "#/components/schemas/BufferAttribute"
            }
          }
        }
      }
    }
  }
}